Simpliflying the logic, at this point communication with server is on sync. Bot is reactive meaning that it will only send the state in two cases: -When finishing a move (so the server can evaluate how effective it was) -When the opponent starts an action (so it can provide a reaction to that attack)
This logic can be improved by adding an idle verification (to know when mac is idle and ask for a move) this requires changes in the server logic as it will need to consider this case. Other option (the best one I think) is send the state on every frame to the server and let the AI decide what an attack but this requires a lot of changes in here (when mac is moving? observe reactions, etc) and can also be CPU intensive. This seems like a fair middle point.