Big Update 4 (#2)
Updates to .NET 4.8
Changes the way action is calculated: I think there was an error on the previous logic as I was probably looking at the memory location and not to the value in the location itself. How it works it explained itself in the code.
Adding Berserker actions: Tracks when an opponent takes a step back and start throwing punches.
Changes in the mainloop: A lot happened so I recommend going through the code, but I simplified it a lot. The logic that I follow ATM is:
Server puts commands on the queue
Client consumes commands and returns a state to the server.
Server puts another command